Factorization of several QTC specific variables
