+++ src/globalvars.h
@@ -120,19 +120,6 @@ extern int three_point;
 extern int dxped;
 extern int addzone;
 
-extern int qsoflags_for_qtc[MAX_QSOS];	// array of flag to log lines of QSOs for QTC's;
-					// this is an array of flags, which marks when a QSO sent as QTC
-extern int nr_qsosflags_for_qtc;	// number of lines in qsoflags_for_qtc[]
-extern int next_qtc_qso;		// the next non-sent QSO, which can be send next as QTC
-extern t_qtclist qtclist;		// the QTC list to send
-extern int nr_qtcsent;
-extern t_qtcreclist qtcreclist;		// the QTC list which received
-
-extern int qtcdirection;				// QTC direction - 1: RECV, 2: SEND
 extern t_pfxnummulti pfxnummulti[MAXPFXNUMMULT];	// array of PFX_NUM_MULTIS parameter
 extern int pfxnummultinr;				// length of array of PFX_NUM_MULTIS parameter
 extern int continentlist_only;				// CONTINENT_LIST_ONLY parameter
-extern t_qtc_ry_line qtc_ry_lines[QTC_RY_LINE_NR];	// when QTC is set, and mode is RTTY, then the modem lines stored this array
-extern int qtc_ry_currline;				// current line of QTC RTTY modem
-extern int qtc_ry_capture;				// enable or disable QTC RTTY capturing
-extern int qtc_ry_copied;				// stores the number of copied lines from QTC RTTY terminal to QTC window
