+++ src/log_recv_qtc_to_disk.c
@@ -32,6 +32,7 @@
 #include <stdlib.h>
 #include "tlf.h"
 #include "lancode.h"
+#include "qtcvars.h"
 
 extern int trx_control;
 extern float freq;
